fixed segfault and started on adjustments
The type table would sometimes segfault because it was passed by value
instead of reference into lower and sema.

Adjustments will eventually be used for type coercian so that the HIR
tree does not need to be modified with casts but instead uses the
hir_info table to store that.